chore: update models with optional info, decimal for monetary

+++ src/models/purchase.py
@@ -1,5 +1,7 @@
 from datetime import datetime
 from typing import Any, Optional
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 from .media import MediaItem
@@ -24,7 +26,7 @@ class SubscriptionBundle(BaseModel):
     id: int
     discount: int
     duration: int
-    price: int
+    price: Decimal
     canBuy: bool
 
 
@@ -61,7 +63,7 @@ class FromUser(BaseModel):
     tipsMax: int
     canEarn: bool
     canAddSubscriber: bool
-    subscribePrice: int
+    subscribePrice: Decimal
     subscriptionBundles: list[SubscriptionBundle]
     isPaywallRequired: bool
     unprofitable: bool
@@ -74,11 +76,11 @@ class FromUser(BaseModel):
     subscribedByExpireDate: datetime
     subscribedByAutoprolong: bool
     subscribedIsExpiredNow: bool
-    currentSubscribePrice: int
+    currentSubscribePrice: Decimal
     subscribedOn: Any
     subscribedOnExpiredNow: Any
     subscribedOnDuration: Any
-    callPrice: int
+    callPrice: Decimal
     lastSeen: str
     canReport: bool
